Use object based-switch, add bench to support the idea
